Generate list of recommended extensions
This is a first take on generating a list of recommended extension
with the view to display in the popup.

The idea is to have a config file for all extensions, and to update it
when an action is undertaken on a recommended extension.

We can add more data per extension like the icon, and maybe a direct
download link to the extension.

Having a direct download link would probably be a good idea only if
it's possible to dynamically change the links (via API?).