Allow proxy to be set per domain instead of the whole browser
This is a major refactor of the way proxy work in the extension.
Instead of a proxy config being used for the whole browser,
each requests are intercepted and can use a specific proxy config
based on arbitrary rules.

The use of the pageAction is removed, because it is non-standard and will eventually be removed by Firefox.

Finally, the interface has been adapted and improved.